+++ src/interface/expense_item.rs
@@ -21,7 +21,7 @@ pub async fn expense_item(
         .await?)
 }
 
-/// Retrieves all Ruddr Expense Item objects, and deserializes it to the corresponding vector of model structs.
+/// Retrieves the first 100 Ruddr Expense Item objects, and deserializes it to the corresponding vector of model structs.
 /// [API Documentation](https://ruddr.readme.io/reference/list-expense-items)
 /// ```ignore
 /// let expense_items = expense_items(&client, Some(types::UUID::try_from("2bdab00d-86fb-46dc-ae05-7cc9c4aedc80").expect("invalid UUID"))).await?;