+++ src/lib.rs
@@ -15,17 +15,17 @@ pub mod interface {
     pub mod utilization;
 }
 pub mod model {
-    pub(crate) mod allocation;
-    pub(crate) mod client;
-    pub(crate) mod cost;
+    pub mod allocation;
+    pub mod client;
+    pub mod cost;
     pub mod enums;
-    pub(crate) mod expense_item;
-    pub(crate) mod expense_report;
-    pub(crate) mod member;
-    pub(crate) mod project;
-    pub(crate) mod role;
-    pub(crate) mod shared;
-    pub(crate) mod time;
+    pub mod expense_item;
+    pub mod expense_report;
+    pub mod member;
+    pub mod project;
+    pub mod role;
+    pub mod shared;
+    pub mod time;
     pub mod types;
-    pub(crate) mod utilization;
+    pub mod utilization;
 }